test: Move the SourceFileCollector fixtures next to its test case
This PR moves the fixtures next to the test case they are used to.
Having fixtures in `tests/phpunit/Fixtures` is useful for re-usability,
but when there is a scenario tailored to a specific use case it is more
convenient to have them next to each other.
