+++ tests/phpunit/FileSystem/SourceFileCollector/Fixtures/case0/outside-symlink.php
